Add files via upload
This version of the react banking app is the framework for creating a three-tier program. Express and MongoDB make the backend and database. Created in React, this version of the banking app is updated and refactored with a data abstract layer file that connects the front-end code to Express and MongoDB.

Note*
This version of the react banking app has not been refracted to keep the state of the user's balance, deposit, and withdrawal. Currently, only the all data and the createUser components are coded to connect with the database.

Future updates will include updated code for the login, deposit, withdrawal, and balance components. Updated features will include verification authorization for the login page and create a user function. The next version, which will be the full three-tier application, may be created in Google's Firebase program.